Feito armazenamento das Tarefas em BD

+++ tcc1/src/main/java/com/tccagil/tcc1/domain/tarefas/TarefasRecord.java
@@ -1,6 +1,6 @@
 package com.tccagil.tcc1.domain.tarefas;
 
-public record TarefasRecord(int trabalhoIdTarefas, String tituloTarefas,
+public record TarefasRecord(Long trabalhoIdTarefas, String tituloTarefas,
         String descricaoTarefas, String prioridadeTarefas, String statusTarefas) {
 
 }
